disable test-rwlock1 with XFAIL
It did not existed on xenial, it is probably a bad test, cannot link to pthread,
but every compiled util is working as expected.

Only sort is using threads, but that is working properly so turning this off is
low risk.

It is added to the generated Makefile.in not to gnulib.mk (and effectively
Makefile.am), because it won't get regenerated so this way we can avoid
dh_autoreconf. A later patch or a dh_autoreconf can break it, but that's
fine as we want to know about that.